# Improve performance of SessionHandler::getSpiderID()