--- a/wcfsetup/install/files/lib/system/cache/builder/SpiderCacheBuilder.class.php
+++ b/wcfsetup/install/files/lib/system/cache/builder/SpiderCacheBuilder.class.php
@@ -23,6 +23,38 @@ class SpiderCacheBuilder extends AbstractCacheBuilder
         $spiderList->sqlOrderBy = "spider.spiderID ASC";
         $spiderList->readObjects();
 
+        if (isset($parameters['fastLookup'])) {
+            $firstCharacter = [];
+            $mapping = [];
+            foreach ($spiderList as $spider) {
+                if (!isset($firstCharacter[$spider->spiderIdentifier[0]])) {
+                    $firstCharacter[$spider->spiderIdentifier[0]] = [];
+                }
+                $firstCharacter[$spider->spiderIdentifier[0]][] = \substr($spider->spiderIdentifier, 1);
+
+                $mapping[$spider->spiderIdentifier] = $spider->spiderID;
+            }
+
+            $regex = '';
+            foreach ($firstCharacter as $char => $spiders) {
+                if ($regex !== '') {
+                    $regex .= '|';
+                }
+                $regex .= \sprintf(
+                    '(?:%s(?:%s))',
+                    \preg_quote($char, '/'),
+                    \implode('|', \array_map(static function ($identifier) {
+                        return \preg_quote($identifier, '/');
+                    }, $spiders))
+                );
+            }
+
+            return [
+                'regex' => "/{$regex}/",
+                'mapping' => $mapping,
+            ];
+        }
+
         return $spiderList->getObjects();
     }
 }